requirements.txt
alibabacloud_cms20190101==2.0.4
alibabacloud_tea_openapi==0.3.3
alibabacloud_tea_util==0.3.5
aliyun_python_sdk_core==2.13.36
aliyun_python_sdk_dds==3.6.6
aliyun_python_sdk_dts==5.1.21
aliyun_python_sdk_ecs==4.24.26
aliyun_python_sdk_r_kvstore==2.20.6
aliyun_python_sdk_rds==2.7.1
aliyun_python_sdk_slb==3.3.17
aliyun_python_sdk_vpc==3.0.17
Flask==2.0.2
oss2==2.15.0
prometheus_client==0.11.0
PyYAML==6.0
Werkzeug==2.0.2
